project/agent-core/requirements.txt
playwright==1.41.2
langgraph==0.2.28
langchain-groq==0.1.9
langchain-core==0.3.10
pydantic==2.9.2
sqlalchemy==2.0.36
python-dotenv==1.0.1
